requirements.txt
langchain-google-vertexai
vertexai
langchain-community
google-cloud-bigquery
sqlalchemy_bigquery
google-cloud-bigquery-storage
sqlalchemy_spanner
fastapi
google-auth

